+++ R/globals.R
@@ -1,4 +1,3 @@
-#' @importFrom utils globalVariables
 #'
 utils::globalVariables(c("name", "iso_a2", "iso_a3", "geometry",
                          "MapFiller", "X", "Y",