+++ R/worldplot.R
@@ -6,7 +6,7 @@
 #' @param ColName Character variable with the name of the variable of interest.
 #' @param CountryName Character variable with the name of the country names column.
 #' @param CountryNameType Character variable with the coding for \code{CountryName}. One of \code{isoa2} (default, standing for ISO 3166-1 alpha-2 code), \code{isoa3}, or \code{name}.
-#' @param rangeVal Limit values that are to be defined for the map.
+#' @param rangeVal Limit values (minimum and maximum) that are to be defined for the map. If not specified, the minimum and maximum are taken, and a message is displayed.
 #' @param longitude Longitude limits. Default is \code{c(-180, 180)} (whole world with crs as EPSG::4326).
 #' @param latitude Latitude limits. Default is \code{c(-90, 90)} (whole world with crs as EPSG::4326).
 #' @param crs Coordinate reference system (EPSG). By default the value is 4326, which corresponds to EPSG::4326 (WGS84)
@@ -15,16 +15,13 @@
 #' @param annote Do you want to plot country labels (ISO 3166-1 alpha-2 code) on the map? Default is set to \code{FALSE}.
 #' @param div Parameter for modifying the elements dimensions in the map. Usually, it does not need to be modified. Default value is 1.
 #' @param palette_option Character string indicating the palette to be used. Available options range between "A" and "H".
+#' @param na_colour The colour to be used for countries with missing information. Default is grey80
+#' @param transform_limits Only if crs is specified and different from 4326. If TRUE (the default) the program expects to receive values of longitude and latitude as in EPSG 4326,
+#'                          (i.e., within -180, +180 for longitude and within -90, +90 for latitude) and automatically updates to the new crs.
+#'                          Set to FALSE if you want to define longitude and latitude limits based on the new crs
 #'
 #' @return a map
 #' @export
-#' @importFrom rnaturalearth ne_countries
-#' @importFrom countrycode countrycode
-#' @importFrom dplyr "%>%" left_join select filter mutate relocate
-#' @importFrom ggplot2 ggplot geom_sf theme labs scale_fill_viridis_c coord_sf xlab ylab ggtitle
-#'                     aes unit element_text element_blank element_rect geom_text
-#' @importFrom sf st_centroid st_coordinates st_union st_as_sf st_transform st_crs
-#' @importFrom ggfx with_shadow
 #'
 #' @examples
 #' data(testdata1b)
@@ -40,7 +37,8 @@ worldplot <- function(data,
                       ColName, CountryName, CountryNameType = "isoa2", rangeVal,
                       longitude = c(-180, 180) ,latitude = c(-90, 90), crs = 4326,
                       title = "", legendTitle = as.character(ColName),
-                      annote = FALSE, div = 1, palette_option = "D") {
+                      annote = FALSE, div = 1, palette_option = "D",
+                      na_colour = "grey80", transform_limits = TRUE) {
 
   world <- ne_countries(scale = 50, continent = NULL, returnclass = "sf")
 
@@ -75,6 +73,12 @@ worldplot <- function(data,
   simdata <- as.data.frame(simdata)
 
   map_df <- left_join(map_df0, simdata, by = "iso_a2")
+  
+
+  if (missing(rangeVal)) {
+    rangeVal = c(range(map_df$MapFiller, na.rm = TRUE))
+  }
+  
 
   wplot <- ggplot(data= map_df) +
     geom_sf(color= 'black', aes(fill= MapFiller)) +
@@ -85,12 +89,35 @@ worldplot <- function(data,
           panel.grid = element_blank(),
           panel.background = element_rect(fill = 'grey95'))+
     labs(fill= legendTitle)+
-    scale_fill_viridis_c(option= palette_option, na.value = 'grey80',direction=1,begin=0.3, limits= rangeVal)+
     coord_sf(xlim= longitude, ylim= latitude, expand= FALSE, label_axes = 'SW') +
     xlab('') + ylab('')+
     ggtitle(title)
+  
+  if (length(palette_option) == 1) {
+    
+    wplot <- wplot +
+      scale_fill_viridis_c(option= palette_option, na.value = na_colour, direction=1,begin=0.3, limits= rangeVal)
+    
+  } else {
+    
+    wplot <- wplot +
+      scale_fill_gradientn(colours = palette_option, na.value = na_colour, limits = rangeVal)
+  }
+  
 
   if (crs != 4326) {
+    
+    if (transform_limits == TRUE) {
+    #Correct longitude and latitude values
+    lim1 <- data.frame(X = longitude, Y = latitude)
+    lim2 <- st_as_sf(lim1, coords=c("X","Y"), crs="EPSG:4326" )
+    lim3 <- st_transform(lim2, crs = st_crs(crs))
+    
+    longitude <- c(st_bbox(lim3)$xmin, st_bbox(lim3)$xmax)
+    latitude <-  c(st_bbox(lim3)$ymin, st_bbox(lim3)$ymax)
+    ##
+    }
+    
     wplot <- wplot +
       coord_sf(xlim= longitude, ylim= latitude, expand= FALSE, label_axes = 'SW',
                crs = st_crs(crs))
@@ -127,8 +154,6 @@ worldplot <- function(data,
       with_shadow(geom_text(data= world_points, aes(x=X, y=Y,label= iso_a2), size= 2/div, color= 'white', fontface= 'bold'),
                   x_offset = 2, y_offset = 2, sigma = 1)
   }
-
-  print(wplot)
 
   return(wplot)