+++ tests/testthat/testCat2.R
@@ -38,7 +38,7 @@ worldplotCat(data = testdata1c,
              Categories = c("Low", "Average", "High"),
              legendTitle = "CAT",
              latitude = c(5350000, 1000000), longitude = c(2500000, 6900000),
-             crs = 3035, annote = TRUE)
+             crs = 3035, annote = TRUE, transform_limits = F)
 
 
 worldplotCat(data = testdata1c,
@@ -50,6 +50,14 @@ worldplotCat(data = testdata1c,
              legendTitle = "CAT",
              latitude = c(5350000, 1000000), longitude = c(2500000, 6900000),
              crs = 3035, annote = TRUE,
-             na.as.category = F)
+             na.as.category = F, transform_limits = F)
 
-
+worldplotCat(data = testdata1c,
+             ColName = "ValCat",
+             CountryName = "iso_a2",
+             CountryNameType = "isoa2",
+             palette_option = c("#C3E2EA", "#58C0D0", "#256C91"),
+             Categories = c("Low", "Average", "High"),
+             legendTitle = "CAT",
+             latitude = c(30,62), longitude = c(-7, 70),
+             crs = 3035, annote = TRUE, transform_limits = T)