+++ NEWS.md
@@ -1,3 +1,11 @@
+# WorldMapR 1.0.1
+
+* Corrected a bug which caused the figures to be plotted twice
+* Default option for range value added
+* Possibility to specify a user-defined gradient continuous variables added
+* Option to modify the color for missing countries added
+* Option to automatically update coordinates limits when using different crs added
+
 # WorldMapR 0.1.3
 
 * Added option to modify the coordinate reference system (crs argument)